requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cx-Oracle==7.3.0
idna==2.8
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
pytest-dependency==0.4.0
pytest-ordering==0.6
pytest-rerunfailures==8.0
requests==2.22.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
